Respect comments in SQL + move code to extensions-base
The issue described in #8221 is in `NessieSparkSqlExtensionsParser#isNessieCommand`, which does not respect SQL comments. The actual fix for that issue is in there. Also slightly updated the antlr grammar.

The rest of this PR is "boring" - unifies the `Base*Exec` into the `*Exec` ones and to the `:nessie_spark-extendions-base` package, leaving just the integration-test hull in the projects for the different Spark versions.

Fixes #8221
