+++ integrations/spark-extensions-base/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/ShowLogCommand.scala
