+++ integrations/spark-extensions-base/src/main/scala/org/apache/spark/sql/execution/datasources/v2/DropReferenceExec.scala
@@ -18,24 +18,44 @@ package org.apache.spark.sql.execution.datasources.v2
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.connector.catalog.CatalogPlugin
+import org.apache.spark.sql.execution.datasources.v2.NessieUtils.unquoteRefName
+import org.apache.spark.unsafe.types.UTF8String
 import org.projectnessie.client.api.NessieApiV1
+import org.projectnessie.error.NessieReferenceNotFoundException
 
-abstract class BaseShowReferenceExec(
+case class DropReferenceExec(
     output: Seq[Attribute],
+    branch: String,
     currentCatalog: CatalogPlugin,
-    catalog: Option[String]
-) extends NessieExec(catalog = catalog, currentCatalog = currentCatalog) {
+    isBranch: Boolean,
+    catalog: Option[String],
+    failOnDrop: Boolean
+) extends NessieExec(catalog = catalog, currentCatalog = currentCatalog)
+    with LeafV2CommandExec {
 
   override protected def runInternal(
       api: NessieApiV1
   ): Seq[InternalRow] = {
-
-    val ref = NessieUtils.getCurrentRef(api, currentCatalog, catalog)
-    // todo have to figure out if this is delta or iceberg and extract the ref accordingly
-    singleRowForRef(ref)
+    val refName = unquoteRefName(branch)
+    try {
+      val hash = api.getReference.refName(refName).get().getHash
+      if (isBranch) {
+        api.deleteBranch().branchName(refName).hash(hash).delete()
+      } else {
+        api.deleteTag().tagName(refName).hash(hash).delete()
+      }
+    } catch {
+      case e: NessieReferenceNotFoundException =>
+        if (failOnDrop) {
+          throw e
+        }
+    }
+    Seq(InternalRow(UTF8String.fromString("OK")))
   }
 
   override def simpleString(maxFields: Int): String = {
-    s"ShowReferenceExec ${catalog.getOrElse(currentCatalog.name())} "
+    s"DropReferenceExec ${catalog.getOrElse(currentCatalog.name())} ${if (isBranch) "BRANCH"
+      else "TAG"} ${unquoteRefName(branch)} "
   }
+
 }
