+++ integrations/spark-extensions-base/src/main/scala/org/apache/spark/sql/execution/datasources/v2/NessieExtendedDataSourceV2Strategy.scala
