+++ integrations/spark-extensions-base/src/main/scala/org/projectnessie/spark/extensions/NessieSparkSessionExtensions.scala
